chore: split app state into smaller files
