+++ packages/ai-chat/src/types/state/PendingUpload.ts
@@ -1,0 +1,61 @@
+/*
+ *  Copyright IBM Corp. 2025, 2026
+ *
+ *  This source code is licensed under the Apache-2.0 license found in the
+ *  LICENSE file in the root directory of this source tree.
+ *
+ *  @license
+ */
+
+import type { StructuredData } from "../messaging/Messages";
+
+/**
+ * The status of a single pending file upload.
+ *
+ * - `UPLOADING` — the upload is in progress; the send button is disabled.
+ * - `COMPLETE` — the upload finished successfully; `contributedData` is populated.
+ * - `ERROR` — the upload failed; `errorMessage` is populated.
+ *
+ * @internal
+ */
+export enum PendingUploadStatus {
+  UPLOADING = "uploading",
+  COMPLETE = "complete",
+  ERROR = "error",
+}
+
+/**
+ * State for a single file that is being (or has been) uploaded via
+ * {@link UploadConfig.onFileUpload}. The chat widget tracks one of these per
+ * in-flight or completed upload and uses them to build the
+ * `pendingStructuredData` attached to the outgoing message.
+ *
+ * @internal
+ */
+export interface PendingUpload {
+  /**
+   * Stable unique identifier for this upload, generated by the widget.
+   */
+  id: string;
+
+  /**
+   * The original `File` object selected by the user.
+   */
+  file: File;
+
+  /**
+   * Current lifecycle status of the upload.
+   */
+  status: PendingUploadStatus;
+
+  /**
+   * The `StructuredData` returned by `onFileUpload` once the upload completes.
+   * Only populated when `status === "complete"`.
+   */
+  contributedData?: StructuredData;
+
+  /**
+   * Human-readable error message when `status === "error"`.
+   */
+  errorMessage?: string;
+}
